requirements.txt
pynfe
google-api-python-client
googlesearch-python
google